feat: mem2reg

+++ testing-example/arithmetic/div.cl
@@ -1,4 +1,4 @@
-__kernel void div(__global float* a, __global float* b, __global float* result) {
+__kernel void div(__global int* a, __global int* b, __global int* result) {
     int i = get_global_id(0);
     result[i] = a[i] / b[i];
 }
