+++ testing-example/arithmetic/div_imm.cl
@@ -1,4 +1,4 @@
-__kernel void div_imm(__global float* a, float imm, __global float* result) {
+__kernel void div_imm(__global int* a, int imm, __global int* result) {
     int i = get_global_id(0);
     result[i] = a[i] / imm;
 }
